evidence: cap evidence to an absolute number (#4780)
The number of evidence that can be committed in a single block is capped by a new evidence parameter called MaxNum

+++ consensus/replay_stubs.go
@@ -50,11 +50,11 @@ type emptyEvidencePool struct{}
 
 var _ sm.EvidencePool = emptyEvidencePool{}
 
-func (emptyEvidencePool) PendingEvidence(int64) []types.Evidence { return nil }
-func (emptyEvidencePool) AddEvidence(types.Evidence) error       { return nil }
-func (emptyEvidencePool) Update(*types.Block, sm.State)          {}
-func (emptyEvidencePool) IsCommitted(types.Evidence) bool        { return false }
-func (emptyEvidencePool) IsPending(types.Evidence) bool          { return false }
+func (emptyEvidencePool) PendingEvidence(uint32) []types.Evidence { return nil }
+func (emptyEvidencePool) AddEvidence(types.Evidence) error        { return nil }
+func (emptyEvidencePool) Update(*types.Block, sm.State)           {}
+func (emptyEvidencePool) IsCommitted(types.Evidence) bool         { return false }
+func (emptyEvidencePool) IsPending(types.Evidence) bool           { return false }
 
 //-----------------------------------------------------------------------------
 // mockProxyApp uses ABCIResponses to give the right results.
